+++ cmd/frontend/internal/cli/http.go
@@ -129,9 +129,6 @@ func secureHeadersMiddleware(next http.Handler) http.Handler {
 		w.Header().Set("X-Content-Type-Options", "nosniff")
 		w.Header().Set("X-XSS-Protection", "1; mode=block")
 		w.Header().Set("X-Frame-Options", "DENY")
-		if hsts := conf.HTTPStrictTransportSecurity(); hsts != "" {
-			w.Header().Set("Strict-Transport-Security", hsts)
-		}
 		// no cache by default
 		w.Header().Set("Cache-Control", "no-cache, max-age=0")