+++ CHANGELOG.md
@@ -23,6 +23,8 @@ All notable changes to Sourcegraph are documented in this file.
 
 ### Removed
 
+- Removed the `httpStrictTransportSecurity` site configuration option. Use [nginx configuration](https://docs.sourcegraph.com/admin/nginx) for this instead.
+
 ### Removed
 
 ## 2.13.2